move `ChannelKeys` to `KeyManager`
Also move `RecoveredChannelKeys` to `LocalKeyManager`, as this is a
specificity of the derivation-based implementation.
